Added missing Blend track on Image clip + remove and add some restrictions and limits on Subtitle and Image clips fields for better management of values
Refactor UIImageClip and UISubtitleClip to use double for position values

- Updated UIImageClip and UISubtitleClip to use double values for x and y position trackpads, enhancing precision.
- Modified width and height trackpads in UIImageClip to accept double values, improving keyframe interpolation.
- Introduced clamping for blend and opacity values in ImageClip to ensure they remain within valid ranges.
- Added functionality to clamp keyframe values in UIKeyframeSheet, enhancing data integrity during edits.

+++ src/main/java/mchorse/bbs_mod/camera/clips/misc/ImageOverlay.java
@@ -19,8 +19,8 @@ public class ImageOverlay
     public float rotation;
     public TextureBlend textureBlend;
 
-    public int x;
-    public int y;
+    public float x;
+    public float y;
     public float width;
     public float height;
     public float anchorX;
@@ -45,7 +45,7 @@ public class ImageOverlay
         this.textureBlend = textureBlend;
     }
 
-    public void updateLayout(int x, int y, float width, float height, float anchorX, float anchorY, float windowX, float windowY, float opacity)
+    public void updateLayout(float x, float y, float width, float height, float anchorX, float anchorY, float windowX, float windowY, float opacity)
     {
         this.x = x;
         this.y = y;
